+++ cmake/unit_tests.cmake
@@ -7,47 +7,47 @@ include(GoogleTest)
 pkg_check_modules(GMOCK gmock)
 
 add_executable(alltests
-	test/tempdir.cc
-	test/file_utils.cc
-	test/msleep.cc
-	test/die_if.cc
-	test/timeout_watcher.cc
-	test/fake_adapter_server.cc
-	test/mock_authenticator.cc
-	test/mock_request.cc
-	test/core/test_container_of.cc
-	test/core/test_string.cc
-	test/core/test_slist.cc
-	test/core/test_path.cc
-	test/core/test_base64.cc
-	test/core/test_status.cc
-	test/core/test_message.cc
-	test/core/test_message_queue.cc
-	test/adapter/test_response_parser.cc
-	test/adapter/test_server.cc
-	test/adapter/test_server_config.cc
-	test/adapter/test_timepoint.cc
-	test/adapter/test_timer.cc
-	test/adapter/test_credentials.cc
-	test/adapter/test_authenticator.cc
-	test/adapter/test_authenticators.cc
-	test/adapter/test_mountpoint.cc
-	test/adapter/test_uuid_mountpoint.cc
-	test/adapter/test_uuid_mountpoint_factory.cc
-	test/adapter/test_fuse_req.cc
-	test/adapter/jsonrpc/test_util.cc
-	test/adapter/jsonrpc/test_is_request.cc
-	test/adapter/jsonrpc/test_request.cc
-	test/adapter/jsonrpc/test_is_response.cc
-	test/adapter/jsonrpc/test_response.cc
-	test/adapter/jsonrpc/test_server.cc
-	test/adapter/jsonrpc/test_proxy.cc
-	test/provider/test_url.cc
-	test/provider/test_static_filesystem.cc
-	test/provider/test_client_protocol.cc
-	test/integration/test_integration.cc
-	test/integration/server.cc
-	test/integration/provider.cc
+	test/webfuse/utils/tempdir.cc
+	test/webfuse/utils/file_utils.cc
+	test/webfuse/utils/msleep.cc
+	test/webfuse/utils/die_if.cc
+	test/webfuse/utils/timeout_watcher.cc
+	test/webfuse/fakes/fake_adapter_server.cc
+	test/webfuse/mocks/mock_authenticator.cc
+	test/webfuse/mocks/mock_request.cc
+	test/webfuse/tests/core/test_container_of.cc
+	test/webfuse/tests/core/test_string.cc
+	test/webfuse/tests/core/test_slist.cc
+	test/webfuse/tests/core/test_path.cc
+	test/webfuse/tests/core/test_base64.cc
+	test/webfuse/tests/core/test_status.cc
+	test/webfuse/tests/core/test_message.cc
+	test/webfuse/tests/core/test_message_queue.cc
+	test/webfuse/tests/adapter/test_response_parser.cc
+	test/webfuse/tests/adapter/test_server.cc
+	test/webfuse/tests/adapter/test_server_config.cc
+	test/webfuse/tests/adapter/test_timepoint.cc
+	test/webfuse/tests/adapter/test_timer.cc
+	test/webfuse/tests/adapter/test_credentials.cc
+	test/webfuse/tests/adapter/test_authenticator.cc
+	test/webfuse/tests/adapter/test_authenticators.cc
+	test/webfuse/tests/adapter/test_mountpoint.cc
+	test/webfuse/tests/adapter/test_uuid_mountpoint.cc
+	test/webfuse/tests/adapter/test_uuid_mountpoint_factory.cc
+	test/webfuse/tests/adapter/test_fuse_req.cc
+	test/webfuse/tests/adapter/jsonrpc/test_util.cc
+	test/webfuse/tests/adapter/jsonrpc/test_is_request.cc
+	test/webfuse/tests/adapter/jsonrpc/test_request.cc
+	test/webfuse/tests/adapter/jsonrpc/test_is_response.cc
+	test/webfuse/tests/adapter/jsonrpc/test_response.cc
+	test/webfuse/tests/adapter/jsonrpc/test_server.cc
+	test/webfuse/tests/adapter/jsonrpc/test_proxy.cc
+	test/webfuse/tests/provider/test_url.cc
+	test/webfuse/tests/provider/test_static_filesystem.cc
+	test/webfuse/tests/provider/test_client_protocol.cc
+	test/webfuse/tests/integration/test_integration.cc
+	test/webfuse/tests/integration/server.cc
+	test/webfuse/tests/integration/provider.cc
 )
 
 target_include_directories(alltests PRIVATE
