+++ test/webfuse/fakes/fake_adapter_server.hpp
