+++ test/webfuse/tests/adapter/jsonrpc/test_is_response.cc
