+++ test/webfuse/utils/file_utils.cc
@@ -1,4 +1,4 @@
-#include "file_utils.hpp"
+#include "webfuse/utils/file_utils.hpp"
 
 #include <sys/types.h>
 #include <sys/stat.h>
