+++ test/webfuse/utils/msleep.cc
@@ -1,4 +1,4 @@
-#include "msleep.hpp"
+#include "webfuse/utils/msleep.hpp"
 #include <ctime>
 
 namespace webfuse_test
