+++ test/webfuse/utils/timeout_watcher.cc
@@ -1,4 +1,4 @@
-#include "timeout_watcher.hpp"
+#include "webfuse/utils/timeout_watcher.hpp"
 #include <stdexcept>
 
 using std::chrono::milliseconds;
